requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.304
langchain-community==0.0.8
openai==0.28.1
faiss-cpu==1.7.4